Fix precision

+++ src/bindings/js/node/src/infer_request.cpp
@@ -174,7 +174,7 @@ Napi::Value InferRequestWrap::infer_dispatch(const Napi::CallbackInfo& info) {
 }
 
 void InferRequestWrap::infer(const Napi::Array& inputs) {
-    for (size_t i = 0; i < inputs.Length(); ++i) {
+    for (uint32_t i = 0; i < inputs.Length(); ++i) {
         auto tensor = value_to_tensor(inputs[i], _infer_request, i);
         _infer_request.set_input_tensor(i, tensor);
     }
@@ -184,7 +184,7 @@ void InferRequestWrap::infer(const Napi::Array& inputs) {
 void InferRequestWrap::infer(const Napi::Object& inputs) {
     const auto& keys = inputs.GetPropertyNames();
 
-    for (size_t i = 0; i < keys.Length(); ++i) {
+    for (uint32_t i = 0; i < keys.Length(); ++i) {
         auto input_name = static_cast<Napi::Value>(keys[i]).ToString().Utf8Value();
         auto value = inputs.Get(input_name);
         auto tensor = value_to_tensor(value, _infer_request, input_name);
